+++ abn/application/test_helper.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/iter8-tools/iter8/base/metrics"
+	"github.com/iter8-tools/iter8/base/summarymetrics"
 	"github.com/stretchr/testify/assert"
 	"sigs.k8s.io/yaml"
 )
@@ -48,7 +48,7 @@ func byteArrayToApplication(name string, data []byte) (*Application, error) {
 	}
 	for _, v := range a.Versions {
 		if v.Metrics == nil {
-			v.Metrics = map[string]*metrics.SummaryMetric{}
+			v.Metrics = map[string]*summarymetrics.SummaryMetric{}
 		}
 	}
