+++ abn/application/version.go
@@ -6,22 +6,22 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/iter8-tools/iter8/base/metrics"
+	"github.com/iter8-tools/iter8/base/summarymetrics"
 )
 
 // Version is information about versions of an application in a Kubernetes cluster
 type Version struct {
 	// List of (summary) metrics for a version
-	Metrics map[string]*metrics.SummaryMetric `json:"metrics" yaml:"metrics"`
+	Metrics map[string]*summarymetrics.SummaryMetric `json:"metrics" yaml:"metrics"`
 }
 
 // GetMetric returns a metric from the list of metrics associated with a version
 // If no metric is present for a given name, a new one is created
-func (v *Version) GetMetric(metric string, allowNew bool) (*metrics.SummaryMetric, bool) {
+func (v *Version) GetMetric(metric string, allowNew bool) (*summarymetrics.SummaryMetric, bool) {
 	m, ok := v.Metrics[metric]
 	if !ok {
 		if allowNew {
-			m := metrics.EmptySummaryMetric()
+			m := summarymetrics.EmptySummaryMetric()
 			v.Metrics[metric] = m
 			return m, true
 		}
