--- base/metrics/doc.go
@@ -1,2 +1,0 @@
-// Package metrics provides a summary metric implementation
-package metrics
