+++ base/summarymetrics/doc.go
@@ -1,0 +1,2 @@
+// Package summarymetrics provides a summary metric implementation
+package summarymetrics
