+++ base/summarymetrics/summary_metric.go
@@ -1,4 +1,4 @@
-package metrics
+package summarymetrics
 
 // summary_metric.go - defines a summary metric object. For space efficiency it is an array of 6 float64.
